initial support for halo 2 original xbox builds. may still have bugs, and some of the beta layouts are almost definitely wrong for the moment, on top of plugins being incomplete, and alpha plugins missing. the buildstring checks are gross hacks, but for now its the best way to manage the minor differences between ogbox and later versions

+++ src/Blamite/Blam/SecondGen/Structures/SecondGenTag.cs
@@ -46,8 +46,8 @@ namespace Blamite.Blam.SecondGen.Structures
 		private void Load(StructureValueCollection values, FileSegmentGroup metaArea, Dictionary<int, ITagGroup> groupsById)
 		{
 			Offset = (int)values.GetInteger("offset");
-			if (Offset > 0)
-				MetaLocation = SegmentPointer.FromPointer(Offset, metaArea);
+			if ((uint)Offset > 0)
+				MetaLocation = SegmentPointer.FromPointer((uint)Offset, metaArea);
 
 			// Load the tag group by looking up the magic value that's stored
 			var groupMagic = (int) values.GetInteger("tag group magic");
